Move error message creation into errors module
